+++ Package.swift
@@ -1,12 +1,12 @@
-// swift-tools-version:5.2
+// swift-tools-version:5.7
 
 import PackageDescription
 
 let package = Package(
     name: "ADCoordinator",
     platforms: [
-        .iOS(.v10),
-        .tvOS(.v10)
+        .iOS(.v13),
+        .tvOS(.v13)
     ],
     products: [
         .library(